Fix CLI handling woes
This commit switches us over to the go-cli library, which is just too
awesome to not use.  The author is a great guy, too.  Additionally,
whitespace between PEM blocks in, and at the end of, the CA certificate
bundle is now ignored.  So now you can use CA bundles exported from
`safe get` or generated by an ERB template, without much hassle.